[Add] Add method to set loan application default status as pending

+++ domain/model/src/main/java/co/com/crediya/model/loanapplication/LoanStatus.java
@@ -9,9 +9,9 @@ public enum LoanStatus {
     APPROVED(2),
     REJECTED(3);
     
-    private final int statusId;
+    private final long statusId;
 
-    LoanStatus(int statusId) {
+    LoanStatus(long statusId) {
         this.statusId = statusId;
     }
